+++ src/app/admin/admin.service.ts
@@ -8,6 +8,7 @@ import {
   AdminSettingsResponse,
   AdminWalletResponse,
   PublisherInfo,
+  RejectedDomainsResponse,
   UserInfo
 } from 'models/settings.model';
 import { environment } from 'environments/environment';
@@ -105,6 +106,14 @@ export class AdminService {
 
   patchPanelPlaceholders(placeholders): Observable<any> {
     return this.http.patch<any>(`${environment.serverUrl}/admin/panel-placeholders`, placeholders);
+  }
+
+  getRejectedDomains(): Observable<RejectedDomainsResponse> {
+    return this.http.get<RejectedDomainsResponse>(`${environment.serverUrl}/admin/rejected-domains`);
+  }
+
+  putRejectedDomains(domains: string[]): Observable<any> {
+    return this.http.put<any>(`${environment.serverUrl}/admin/rejected-domains`, {domains});
   }
 
   getTargetingCriteria(): Observable<TargetingOption[]> {