+++ src/app/models/app-state.model.ts
@@ -61,6 +61,7 @@ interface AdminState {
   settings: AdminSettings;
   wallet: AdminWallet;
   termsAndPrivacy: TermsAndPrivacy;
+  rejectedDomains: string[];
   license: License | null;
   index: Index | null;
   panelBlockade: boolean;